simphony: deal with eic-opticks rename
